% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umclassr.R
\name{get_course_description}
\alias{get_course_description}
\title{This function gets a course description}
\usage{
get_course_description(term, sub, catnum,
  access_token = get_umsched_access_token())
}
\arguments{
\item{term}{term code}

\item{sub}{subject}

\item{catnum}{course number}

\item{access}{token, retrieves a token if one isn't provided}
}
\description{
This function gets a course description
}
\examples{
\dontrun{
get_course_description('2210', 'IOE', '436')
}
}
\keyword{course}
\keyword{description}
