% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umclassr.R
\name{get_schools}
\alias{get_schools}
\title{This function gets all schools offering courses for a term}
\usage{
get_schools(term, access_token = get_umsched_access_token())
}
\arguments{
\item{term}{term code}

\item{access_token}{access token, retrieves a token if one isn't provided}
}
\description{
This function gets all schools offering courses for a term
}
\examples{
\dontrun{
get_schools('2210')
}
}
\keyword{description}
\keyword{school}
