% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umclassr.R
\name{get_umsched_access_token}
\alias{get_umsched_access_token}
\title{This function creates an access token. Its used by the subsequent functions for API authorization.}
\usage{
get_umsched_access_token(client_id = Sys.getenv("UMSCHED_CLIENT_ID"),
  client_secret = Sys.getenv("UMSCHED_CLIENT_SECRET"))
}
\arguments{
\item{client_id}{Defaults to System Environment variable "UMSCHED_CLIENT_ID"}

\item{client_secret}{Defaults to System Environment variable "UMSCHED_CLIENT_SECRET"}
}
\description{
This function creates an access token. Its used by the subsequent functions for API authorization.
}
\examples{
\dontrun{
get_umsched_access_token()
}
}
\keyword{auth}
